moved code in common module into led_finder.py
